Refactor login form and add OAuth login
